some1

+++ src/main/java/ru/pavlov/Factorial.java
@@ -5,12 +5,10 @@ public class Factorial {
         if (n < 0) {
             throw new IllegalArgumentException("Отрицательное число недопустимо");
         }
-
         if (n > 20) {
             throw new NumberTooLargeException("Не доступно, для n > 20");
         }
-
-        int result = 1;
+        long result = 1;
         for (int i = 2; i <= n; i++) {
             result *= i;
         }
